test(extension): cleanup 2
