+++ packages/e2e-tests/src/assert/multidelegationPageAssert.ts
@@ -30,7 +30,24 @@ class MultidelegationPageAssert {
   };
 
   assertSeeDelegationCardDetailsInfo = async () => {
-    await MultidelegationPage.delegationCardLooksCorrectly();
+    expect(await MultidelegationPage.delegationCardBalanceLabel.getText()).to.equal(
+      await t('overview.delegationCard.label.balance', 'staking')
+    );
+    const adaValue = Number((await MultidelegationPage.delegationCardBalanceValue.getText()).split(' ')[0]);
+    expect(adaValue).to.be.greaterThan(0);
+    expect(await MultidelegationPage.delegationCardPoolsLabel.getText()).to.equal(
+      await t('overview.delegationCard.label.pools', 'staking')
+    );
+    const poolsCount = Number(await MultidelegationPage.delegationCardPoolsValue.getText());
+    expect(poolsCount).to.be.greaterThan(0);
+    expect(await MultidelegationPage.delegationCardStatusLabel.getText()).to.equal(
+      await t('overview.delegationCard.label.status', 'staking')
+    );
+    const statusValue = await MultidelegationPage.delegationCardStatusValue.getText();
+    poolsCount === 1
+      ? expect(statusValue).to.equal(await t('overview.delegationCard.statuses.simpleDelegation', 'staking'))
+      : expect(statusValue).to.equal(await t('overview.delegationCard.statuses.multiDelegation', 'staking'));
+    expect(await MultidelegationPage.delegationCardChartSlices.length).to.equal(poolsCount);
   };
 
   assertSeeDelegatedPoolCardsPopup = async () => {
